requirements.txt
torch
torchvision
pillow
tqdm
flask
flask-cors
facenet-pytorch
